+++ src/statistics/stats_visualization.py
@@ -1,9 +1,13 @@
+from typing import Final
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
+import seaborn as sns
 from pandas import DataFrame
 
 
 class StatsVisualization:
+    STYLE_PALETTE: Final[str] = "Dark2"
+
     def bar_daily_worked_minutes(self, title: str, stats: DataFrame):
         """Creates a bar chart showing the daily worked minutes, breaks and the daily target line.
 
@@ -11,32 +15,47 @@ class StatsVisualization:
             stats (DataFrame): Statistics holding information about daily worked minutes and breaks.
         """
         fig, ax = plt.subplots(num="PyTimeTrack")
+
+        # Transform data from wide to long format for visualization
+        df_stats_vis = stats.reset_index()
+        df_stats_vis.drop("default_break_minutes", axis=1, inplace=True)
+        df_stats_vis = df_stats_vis.rename(
+            columns={
+                "total_work_minutes": "Work",
+                "total_break_minutes": "Break",
+                "total_work_without_break": "Pure Work",
+                "target_work_minutes": "Target",
+            }
+        )
+        df_stats_vis = df_stats_vis.melt(
+            id_vars="day",
+            value_vars=[
+                "Work",
+                "Break",
+                "Pure Work",
+                "Target",
+            ],
+        )
+        df_stats_vis = df_stats_vis.rename(columns={"variable": "type"})
+
+        # Adapt plot
         ax.set_title(title)
         fig.autofmt_xdate()
         xfmt = mdates.DateFormatter("%d.%m")
         ax.xaxis.set_major_formatter(xfmt)
         ax.set_xlabel("day")
         ax.set_ylabel("minutes")
+        plt.grid()
 
-        x_axis = mdates.date2num(stats["day"])
-        x_width = 0.3
-
-        ax.bar(
-            x_axis - x_width,
-            stats["total_work_minutes"],
-            width=x_width,
+        sns.lineplot(
+            data=df_stats_vis,
+            x="day",
+            y="value",
+            hue="type",
+            ax=ax,
+            style="type",
+            markers=True,
+            dashes=True,
+            palette=self.STYLE_PALETTE,
         )
-        ax.bar(
-            x_axis,
-            stats["total_break_minutes"],
-            width=x_width,
-        )
-        ax.bar(
-            x_axis + x_width,
-            stats["total_work_without_break"],
-            width=x_width,
-        )
-
-        plt.legend(["Work", "Break", "Pure Work", "Target Work"])
-        plt.axhline(y=stats["target_work_minutes"][0], color="red")
         plt.show()